requirements.txt
telethon
PyQT5
pillow
pytesseract